Use an ES module loader that maps yjs imports to src/index.js
Because of transitive dependencies, a Y.js dependency is installed under
node_modules and is imported by @y/protocols package that's used in some
Y.js tests. This import causes two separate Y.js instances to be
loaded, which in turn wreaks havoc, as a lot of `instanceof` or other
checks in the Y.js core code start to fail.